requirements.txt
requests==2.20.1
python-telegram-bot==12.2.0